Refactor: Rename fragments for clarity and consistency
This commit introduces two main refactoring changes to improve the clarity and consistency of fragment naming within the application.

1.  **Rename `TakeExamFragment` to `ExamTakingFragment`:**
    *   Renamed the fragment class and its corresponding file (`.kt`).
    *   Renamed the associated layout file from `fragment_take_exam.xml` to `fragment_exam_taking.xml`.
    *   Updated all references, including the generated data binding class, in `CourseStepFragment`, `SubmissionsAdapter`, and the fragment itself.

2.  **Rename `LeadersFragment` to `CommunityLeadersFragment`:**
    *   Renamed the fragment class and its corresponding file to better reflect its role within the community section.
    *   Updated the reference in `CommunityPagerAdapter` to use the new class name.
