+++ app/build.gradle
@@ -10,8 +10,8 @@ android {
         applicationId "org.ole.planet.myplanet"
         minSdk = 26
         targetSdk = 36
-        versionCode = 4219
-        versionName = "0.42.19"
+        versionCode = 4216
+        versionName = "0.42.16"
         ndkVersion = '26.3.11579264'
         vectorDrawables.useSupportLibrary = true
     }
