+++ app/src/lite/AndroidManifest.xml
@@ -77,10 +77,10 @@
             android:name=".ui.courses.CourseProgressActivity"
             android:theme="@style/AppTheme" />
         <activity
-            android:name=".ui.health.AddMyHealthActivity"
+            android:name=".ui.myhealth.AddMyHealthActivity"
             android:theme="@style/AppTheme" />
         <activity
-            android:name=".ui.health.AddExaminationActivity"
+            android:name=".ui.myhealth.AddExaminationActivity"
             android:theme="@style/AppTheme" />
         <activity android:name=".ui.maps.OfflineMapActivity" />
         <activity
