+++ app/src/main/java/org/ole/planet/myplanet/ui/courses/CoursesFragment.kt
@@ -328,7 +328,7 @@ class CoursesFragment : BaseRecyclerFragment<RealmMyCourse?>(), OnCourseItemSele
             requireView().findViewById<View>(R.id.fabMyProgress).apply {
                 visibility = View.VISIBLE
                 setOnClickListener {
-                    val progressFragment = CoursesProgressFragment().apply {
+                    val progressFragment = ProgressFragment().apply {
                         arguments = Bundle().apply {
                             putBoolean("isMyCourseLib", true)
                         }
