+++ app/src/main/res/layout/fragment_progress.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.courses.CoursesProgressFragment">
+    tools:context=".ui.courses.ProgressFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_myprogress"
